Delay echoing until end of widget function, and remove outer container from widget.
The outer container on a menu will have menu classes which might already be in use
elsewhere (they were on my site), so that's a no-go.

As for the output, by delaying it we can:
(a) include before / after widget content for a variety of situations,
(b) still avoid outputting any code if there will be no menu inside, and
(c) maintain the code in a slightly more DRY fashion.